requirements.txt
# 核心依赖
flask>=2.3.0
aiohttp>=3.8.0
asyncio>=3.4.3
PyYAML>=6.0
requests>=2.31.0

# 数据库
sqlalchemy>=2.0.0

# 测试工具
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 工具库
python-dotenv>=1.0.0
colorama>=0.4.6
